improved security, added token reset after password change

+++ chatServ/routes/users.js
@@ -4,6 +4,7 @@ const User = require('../models/user');
 const path = require('path');
 const fs = require('fs');
 const crypto = require('crypto');
+const authPayload = require('../modules/tokenPayload');
 
 const multer = require('multer');
 const storage = multer.diskStorage({
@@ -18,26 +19,29 @@ const upload = multer({
     storage: storage,
     limits: {fileSize: 5 * 1024 * 1024}
 });
-//
+
 router.put('/editProfile', upload.single('avatar'), async (request, response) => {
     let oldProfile = await User.findOne({_id: request.user._id});
     if(request.file) {
         request.body.avatar = `avatars/${request.file.filename}`;
-        fs.unlink(`public/${oldProfile.avatar}`, err => console.log(err));
+        if(oldProfile.avatar !== 'avatars/default.jpg') {
+            fs.unlink(`public/${oldProfile.avatar}`, err => console.log(err));
+        }
     }
     // else {
     //     delete request.body.avatar;
     // }
-    // if(request.body.password) {
-    //     request.body.session_hash = crypto.randomBytes(20).toString('hex');
-    // }
+    if(request.body.password) {
+        request.body.session_hash = crypto.randomBytes(20).toString('hex');
+    }
     let newProfile = await User.findOneAndUpdate(
         {_id: request.user._id},
         request.body,
         {new: true, runValidators: true});
-
     if(newProfile) {
-        response.send(newProfile);
+        let session_hash_data = await User.findOne(request.body).select('session_hash');
+        newProfile.session_hash = session_hash_data.session_hash;
+        response.send(authPayload(newProfile));
     } else {
         response.sendStatus(404);
     }
